+++ server/src/database/commands/data-seed-tenant.command.ts
@@ -3,7 +3,7 @@ import { InjectDataSource } from '@nestjs/typeorm';
 import { Command, CommandRunner } from 'nest-commander';
 import { DataSource } from 'typeorm';
 
-import { DataSourceMetadataService } from 'src/metadata/data-source-metadata/data-source-metadata.service';
+import { DataSourceService } from 'src/metadata/data-source/data-source.service';
 import { TenantMigrationService } from 'src/metadata/tenant-migration/tenant-migration.service';
 import { TenantMigrationRunnerService } from 'src/tenant-migration-runner/tenant-migration-runner.service';
 import { seedCompanies } from 'src/database/typeorm-seeds/tenant/companies';
@@ -25,7 +25,7 @@ export class DataSeedTenantCommand extends CommandRunner {
   constructor(
     @InjectDataSource('metadata')
     private readonly metadataDataSource: DataSource,
-    private readonly dataSourceMetadataService: DataSourceMetadataService,
+    private readonly dataSourceMetadataService: DataSourceService,
     private readonly typeORMService: TypeORMService,
     private readonly tenantMigrationService: TenantMigrationService,
     private readonly migrationRunnerService: TenantMigrationRunnerService,
