+++ server/src/database/commands/database-command.module.ts
@@ -7,7 +7,7 @@ import { CompanyModule } from 'src/core/company/company.module';
 import { PersonModule } from 'src/core/person/person.module';
 import { PrismaModule } from 'src/database/prisma.module';
 import { TenantManagerModule } from 'src/tenant-manager/tenant-manager.module';
-import { DataSourceMetadataModule } from 'src/metadata/data-source-metadata/data-source-metadata.module';
+import { DataSourceModule } from 'src/metadata/data-source/data-source.module';
 import { TenantMigrationModule } from 'src/metadata/tenant-migration/tenant-migration.module';
 import { TenantMigrationRunnerModule } from 'src/tenant-migration-runner/tenant-migration-runner.module';
 import { TypeORMModule } from 'src/database/typeorm/typeorm.module';
@@ -22,7 +22,7 @@ import { DataSeedTenantCommand } from './data-seed-tenant.command';
     PersonModule,
     TenantManagerModule,
     PrismaModule,
-    DataSourceMetadataModule,
+    DataSourceModule,
     TypeORMModule,
     TenantMigrationModule,
     TenantMigrationRunnerModule,
