+++ server/src/metadata/object-metadata/object-metadata.module.ts
@@ -7,7 +7,7 @@ import {
 import { NestjsQueryTypeOrmModule } from '@ptc-org/nestjs-query-typeorm';
 import { SortDirection } from '@ptc-org/nestjs-query-core';
 
-import { DataSourceMetadataModule } from 'src/metadata/data-source-metadata/data-source-metadata.module';
+import { DataSourceModule } from 'src/metadata/data-source/data-source.module';
 import { TenantMigrationRunnerModule } from 'src/tenant-migration-runner/tenant-migration-runner.module';
 import { TenantMigrationModule } from 'src/metadata/tenant-migration/tenant-migration.module';
 import { JwtAuthGuard } from 'src/guards/jwt.auth.guard';
@@ -24,7 +24,7 @@ import { ObjectMetadataDTO } from './dtos/object-metadata.dto';
     NestjsQueryGraphQLModule.forFeature({
       imports: [
         NestjsQueryTypeOrmModule.forFeature([ObjectMetadataEntity], 'metadata'),
-        DataSourceMetadataModule,
+        DataSourceModule,
         TenantMigrationModule,
         TenantMigrationRunnerModule,
       ],
