+++ server/src/tenant-manager/commands/sync-tenant-metadata.command.ts
@@ -1,6 +1,6 @@
 import { Command, CommandRunner, Option } from 'nest-commander';
 
-import { DataSourceMetadataService } from 'src/metadata/data-source-metadata/data-source-metadata.service';
+import { DataSourceService } from 'src/metadata/data-source/data-source.service';
 import { TenantManagerService } from 'src/tenant-manager/tenant-manager.service';
 
 // TODO: implement dry-run
@@ -15,7 +15,7 @@ interface RunTenantMigrationsOptions {
 export class SyncTenantMetadataCommand extends CommandRunner {
   constructor(
     private readonly tenantManagerService: TenantManagerService,
-    private readonly dataSourceMetadataService: DataSourceMetadataService,
+    private readonly dataSourceMetadataService: DataSourceService,
   ) {
     super();
   }
