+++ server/src/tenant/tenant.service.spec.ts
@@ -1,6 +1,6 @@
 import { Test, TestingModule } from '@nestjs/testing';
 
-import { DataSourceMetadataService } from 'src/metadata/data-source-metadata/data-source-metadata.service';
+import { DataSourceService } from 'src/metadata/data-source/data-source.service';
 import { ObjectMetadataService } from 'src/metadata/object-metadata/object-metadata.service';
 
 import { TenantService } from './tenant.service';
@@ -16,7 +16,7 @@ describe('TenantService', () => {
       providers: [
         TenantService,
         {
-          provide: DataSourceMetadataService,
+          provide: DataSourceService,
           useValue: {},
         },
         {
